requirements.txt
numpy
pandas
opencv-python
Pillow
matplotlib
requests
openpyxl